requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
colorama>=0.4.6
psutil>=5.9.0
matplotlib>=3.7.0
folium>=0.14.0
requests>=2.31.0